[cleanup] erefactor/EclipseJdt - Remove unnecessary casts
EclipseJdt cleanup 'RemoveUnnecessaryCast' applied by erefactor.

For EclipseJdt see https://www.eclipse.org/eclipse/news/4.18/jdt.php
For erefactor see https://github.com/cal101/erefactor

+++ modules/openapi-generator/src/main/java/org/openapitools/codegen/languages/JavaCXFExtServerCodegen.java
@@ -708,8 +708,8 @@ public class JavaCXFExtServerCodegen extends JavaCXFServerCodegen implements CXF
             // NOTE: use double to hold float values, to avoid numeric overflow.
             double min = var == null || var.minimum == null ? -Float.MAX_VALUE : Float.parseFloat(var.minimum);
             double max = var == null || var.maximum == null ? Float.MAX_VALUE : Float.parseFloat(var.maximum);
-            double exclusiveMin = (double) (var != null && var.exclusiveMinimum ? 1 : 0);
-            double inclusiveMax = (double) (var == null || !var.exclusiveMaximum ? 1 : 0);
+            double exclusiveMin = var != null && var.exclusiveMinimum ? 1 : 0;
+            double inclusiveMax = var == null || !var.exclusiveMaximum ? 1 : 0;
             float randomFloat = (float) (min + exclusiveMin
                     + ((max + inclusiveMax - min - exclusiveMin) * Math.random()));
